Don't test for 127.0.0.1 when destroying services
- 'docker-compose stop' will work fine and not do anything if that docker container isn't running

Resolves #44

Removing test for DOCKER_SERVER env var

Don't error if docker-compose doesn't exist, just skip the command

+++ local/server/destroy.sh
@@ -1,8 +1,11 @@
 #!/bin/bash
 
-PID=$(cat app.pid)
+if [ -f 'app.pid' ]; then
 
-if ps -p  $PID > /dev/null; then
-  echo "Killing process PID [$PID]" >&2
-  kill -9 $PID > /dev/null 2>&1
+  PID=$(cat app.pid)
+  if ps -p $PID > /dev/null; then
+    echo "Killing process PID [$PID]" >&2
+    kill -9 $PID > /dev/null 2>&1
+  fi
+
 fi
